+++ notification_dispatcher/lib/src/typedefs.dart
@@ -1,4 +1,5 @@
-import 'notification_dispatcher.dart';
-
 /// Callback signature when receiving a notification.
-typedef NotificationCallback = void Function(NotificationMessage message);
+typedef NotificationCallback = void Function(NotificationMessage? message);
+
+/// The message being posted/received.
+typedef NotificationMessage = Map<String, dynamic>;